+++ test/fixtures/script_example/outputs.tf
@@ -14,17 +14,7 @@
  * limitations under the License.
  */
 
-output "create_cmd_bin" {
-  description = "The full bin path & command used on create"
-  value       = local.create_cmd_bin
-}
-
-output "destroy_cmd_bin" {
-  description = "The full bin path & command used on destroy"
-  value       = local.destroy_cmd_bin
-}
-
-output "bin_dir" {
-  description = "The full bin path of the modules executables"
-  value       = local.gcloud_bin_path
+output "project_id" {
+  description = "The ID of the project in which resources are provisioned."
+  value       = var.project_id
 }