+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- `create_cmd_entrypoint` and `destroy_cmd_entrypoint` variables can now be set to a custom script to run. Prior to running the command the module will prepend the module's bin directory to `PATH`. [#9]
+- `create_cmd_triggers` can be set to trigger the `create_cmd` local-exec again. [#9]
+
 ## [0.2.0] - 2019-12-18
 
 ### Changed
@@ -24,3 +29,5 @@ and this project adheres to
 [Unreleased]: https://github.com/terraform-google-modules/terraform-google-gcloud/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/terraform-google-modules/terraform-google-gcloud/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/terraform-google-modules/terraform-google-gcloud/releases/tag/v0.1.0
+
+[#9]: https://github.com/terraform-google-modules/terraform-google-gcloud/pull/9